+++ sdk/deepbook-v3/test/client.test.ts
@@ -1,13 +1,10 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import path from 'path';
 import { beforeAll, describe, expect, test } from 'vitest';
-
 import { DeepBookClient, DeepBookConfig } from '../src';
 import { CoinMap } from '../src/utils/constants';
-import { publishPackage, setupSuiClient, TestToolbox } from './setup';
-import { writeToml } from './helper/toml';
+import { publishCoins, publishDeepBook, setupSuiClient, TestToolbox } from './setup';
 
 let toolbox!: TestToolbox;
 let coins: CoinMap;
@@ -17,47 +14,11 @@ let deepbookAdminCap: string;
 
 beforeAll(async () => {
 	toolbox = await setupSuiClient();
-	const tokenSourcesPath = path.join(__dirname, 'data/token');
-	writeToml(tokenSourcesPath, '0x0', 'token');
-	let deepRes = await publishPackage(tokenSourcesPath, toolbox);
-	writeToml(tokenSourcesPath, deepRes.packageId, 'token');
-
-	const usdcSourcePath = path.join(__dirname, 'data/usdc');
-	const usdcRes = await publishPackage(usdcSourcePath, toolbox);
-
-	const spamSourcePath = path.join(__dirname, 'data/spam');
-	const spamRes = await publishPackage(spamSourcePath, toolbox);
-
-	const deepbookSourcesPath = path.join(__dirname, 'data/deepbook');
-	let deepbookRes = await publishPackage(deepbookSourcesPath, toolbox);
-
-	deepbookPackageId = deepbookRes.packageId;
-	// @ts-ignore
-	deepbookRegistryId = deepbookRes.publishTxn.objectChanges?.find((change) => {
-		return change.type === 'created' && change.objectType.includes('Registry') && !change.objectType.includes('Inner')
-	})?.['objectId'];
-
-	// @ts-ignore
-	deepbookAdminCap = deepbookRes.publishTxn.objectChanges?.find((change) => {
-		return change.type === 'created' && change.objectType.includes('DeepbookAdminCap');
-	})?.['objectId'];
-	coins = {
-		DEEP: {
-			address: deepRes.packageId,
-			type: `${deepRes.packageId}::deep::DEEP`,
-			scalar: 1000000,
-		},
-		USDC: {
-			address: usdcRes.packageId,
-			type: `${usdcRes.packageId}::usdc::USDC`,
-			scalar: 1000000,
-		},
-		SPAM: {
-			address: spamRes.packageId,
-			type: `${spamRes.packageId}::spam::SPAM`,
-			scalar: 1000000,
-		},
-	};
+    coins = await publishCoins(toolbox);
+    const res = await publishDeepBook(toolbox);
+    deepbookPackageId = res.deepbookPackageId;
+    deepbookRegistryId = res.deepbookRegistryId;
+    deepbookAdminCap = res.deepbookAdminCap;
 });
 
 describe('DeepbookClient', () => {
